+++ Include/Acidanthera/Library/OcMacInfoLib.h
@@ -15,6 +15,16 @@
 #ifndef OC_MAC_INFO_LIB_H
 #define OC_MAC_INFO_LIB_H
 
+//
+// Maximum characters for valid Mac-like product name.
+//
+#define OC_OEM_NAME_MAX 48
+
+//
+// Maximum characters for valid Mac-like serial name.
+//
+#define OC_OEM_SERIAL_MAX 16
+
 typedef struct MAC_INFO_DATA_SMBIOS_ {
   //
   // Type 0
@@ -71,6 +81,12 @@ typedef struct MAC_INFO_DATA_DATAHUB_ {
   CONST UINT8   *SmcPlatform;
 } MAC_INFO_DATA_DATAHUB;
 
+typedef struct MAC_INFO_DATA_OEM_ {
+  CHAR8         SystemSerialNumber[OC_OEM_SERIAL_MAX];
+  CHAR8         BoardSerialNumber[OC_OEM_SERIAL_MAX];
+  EFI_GUID      SystemUuid;
+} MAC_INFO_DATA_OEM;
+
 typedef struct MAC_INFO_DATA_ {
   //
   // DataHub data.
@@ -80,6 +96,10 @@ typedef struct MAC_INFO_DATA_ {
   // SMBIOS data.
   //
   MAC_INFO_DATA_SMBIOS  Smbios;
+  //
+  // Serial data.
+  //
+  MAC_INFO_DATA_OEM     Oem;
 } MAC_INFO_DATA;
 
 /**