+++ Changelog.md
@@ -2,6 +2,7 @@ OpenCore Changelog
 ==================
 #### v0.6.7
 - Fixed ocvalidate return code to be non-zero when issues are found
+- Added `OEM` values to `PlatformInfo` in `Automatic` mode
 
 #### v0.6.6
 - Added keyboard and pointer entry scroll support in OpenCanopy